chore(points): update more coming icon
